# openapi.yaml
openapi: 3.0.3
info:
  title: Activity Tracker REST API
  description: |
    REST API for the Activity Tracker Minecraft plugin. This API allows external applications 
    to retrieve player activity statistics, leaderboards, and detailed player information.
  version: 1.2.0
  contact:
    name: Activity Tracker
    url: https://github.com/Dans-Plugins/Activity-Tracker
  license:
    name: GPL-3.0
    url: https://github.com/Dans-Plugins/Activity-Tracker/blob/master/LICENSE
servers:
  - url: http://localhost:8080
    description: Local development server (default port)
...
paths:
...
  /api/players/{uuid}:
    get:
      summary: Get player activity details
      description: Retrieve detailed activity information for a specific player by their UUID
      operationId: getPlayerActivity
      tags:
        - Players
      parameters:
        - name: uuid
          in: path
          required: true
          description: The UUID of the player
          schema:
            type: string
            format: uuid
            example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        '200':
          description: Player activity details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerActivityResponse'
        '400':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    PlayerActivityResponse:
      type: object
      properties:
        playerUuid:
          type: string
          format: uuid
          description: UUID of the player
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        playerName:
          type: string
          description: Name of the player
          example: "Steve"
        totalLogins:
          type: integer
          description: Total number of times the player has logged in
          example: 87
        totalHoursPlayed:
          type: number
          format: double
          description: Total hours played by the player
          example: 123.45
        currentlyOnline:
          type: boolean
          description: Whether the player is currently online
          example: true
        firstLogin:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the player's first login
          example: "2023-01-15T14:30:00"
        lastLogin:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the player's most recent login
          example: "2023-10-26T10:15:00"
        lastLogout:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the player's most recent logout
          example: "2023-10-26T12:45:00"
        hoursSinceLogin:
          type: number
          format: double
          nullable: true
          description: Hours since the player logged in (only present if currently online)
          example: 2.5
        hoursSinceLogout:
          type: number
          format: double
          nullable: true
          description: Hours since the player logged out (only present if currently offline)
          example: 48.3
      required:
        - playerUuid
        - playerName
        - totalLogins
        - totalHoursPlayed
        - currentlyOnline
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Player not found"
      required:
        - error